solutions/audio/requirements.txt
# Gradio to create the application, see https://pypi.org/project/gradio/
gradio==5.5.0
# NVIDIA client to use AIEndpoints, see https://pypi.org/project/nvidia-riva-client/, https://github.com/nvidia-riva/python-clients
nvidia-riva-client==2.17.0 
# Pydub to manipulate audio files, see https://pypi.org/project/pydub/
pydub==0.25.1
# Numpy to do sientific computing with Python
numpy==2.1.3
# Manage http requests, see https://pypi.org/project/requests/
requests==2.32.3
